Reapply "pruebas unitarias a los metodos Crud de JPARepository"
This reverts commit 644fce9b65f015816489c4a92f2ffcdb6198bb82.

+++ src/main/java/com/example/demo/repository/SugerenciaRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.modelo.Sugerencia;
 
-public interface SugerenciaRepository extends JpaRepository<Sugerencia, Integer>{
-    
+import java.util.List;
+
+public interface SugerenciaRepository extends JpaRepository<Sugerencia, Long>{
+
 }
